requirements.txt
requests==2.32.3
pandas==2.2.3
fredapi==0.5.2
python-dotenv==1.2.2
